+++ src/DeckPlayer/Action.hs
@@ -21,19 +21,20 @@ import DeckPlayer.Assets
 import DeckFormat.DeckFormat
 import DeckFormat.Structure (DeckDirectory(..), titleCardName, deckLookup)
 import Control.Monad (foldM)
+import Data.Word (Word32)
 
 {- | A wrapper for performing a series of actions, accumulating the state. 
 
 -}
-handleActions :: Renderer -> [Action] -> CardDeck -> DeckState -> IO DeckState
-handleActions renderer actions deck deckState =
-    foldM (\state action -> handleAction renderer action deck state) deckState actions
+handleActions :: Renderer -> [Action] -> CardDeck -> Word32 -> DeckState -> IO DeckState
+handleActions renderer actions deck currentTicks deckState =
+    foldM (\state action -> handleAction renderer action deck currentTicks state) deckState actions
 
 {- | Handle/perform an 'Action', possibly conditionally.
 
 -}
-handleAction :: Renderer -> Action -> CardDeck -> DeckState -> IO DeckState
-handleAction renderer actionOrConditional deck deckState =
+handleAction :: Renderer -> Action -> CardDeck -> Word32 -> DeckState -> IO DeckState
+handleAction renderer actionOrConditional deck currentTicks deckState =
     case actionOrConditional of
         (ConditionalAction (Conditional conditionals actionOnTrue actionOnFalse)) -> do
             -- In the future there will be more conditionals and we'll basically do an AND match on them.
@@ -49,15 +50,15 @@ handleAction renderer actionOrConditional deck deckState =
                 then do
                     case actionOnTrue of
                         Just actionOnTrue' ->
-                            handleAction' actionOnTrue' deck deckState
+                            handleAction' actionOnTrue' deck currentTicks deckState
                         Nothing ->
                             return deckState
                 else do
                     case actionOnFalse of
-                        Just actionOnFalse' -> handleAction' actionOnFalse' deck deckState
+                        Just actionOnFalse' -> handleAction' actionOnFalse' deck currentTicks deckState
                         Nothing -> return deckState
 
-        action -> handleAction' action deck deckState
+        action -> handleAction' action deck currentTicks deckState
    where
     performFlagConditional (ActionFlag requireThisFlag requireThisFlagValue) =
         let
@@ -65,8 +66,8 @@ handleAction renderer actionOrConditional deck deckState =
         in
             flagActualValue == requireThisFlagValue
 
-    handleAction' :: Action -> CardDeck -> DeckState -> IO DeckState
-    handleAction' action deck deckState =
+    handleAction' :: Action -> CardDeck -> Word32 -> DeckState -> IO DeckState
+    handleAction' action deck currentTicks deckState =
         case action of
             (ActionTypeLink (ActionLink linkCard')) -> do
                 let
@@ -80,11 +81,11 @@ handleAction renderer actionOrConditional deck deckState =
                 return deckState
             (ActionTypeScript script) -> do
                 let (AssetScript scriptAsset) = assetLookupError DirectoryScripts (script ^. actionScriptAsset) (deckState ^. deckStateAssetRegistry)
-                runLuaScript renderer deck deckState scriptAsset
+                runLuaScript renderer (fromIntegral currentTicks :: Int) deck deckState scriptAsset
             (ActionTypeFlag (ActionFlag flagName' flagValue')) -> do
                 return $ deckStateFlags .~ setFlag (deckState ^. deckStateFlags) flagName' flagValue' $ deckState
             conditionalAction@(ConditionalAction _) ->
-                handleAction renderer conditionalAction deck deckState
+                handleAction renderer conditionalAction deck currentTicks deckState
 
 {- | Checks if the cordinate '(Cint, Cint)' is inside of 'CardObject'.